Services (#9)
* update services & controllers with request handling

* update services with kitchen to hall communication setup

* update services & controller

* update service concurrency

* update order delivery logic & rating system

+++ src/components/types/order/delivery.go
@@ -10,6 +10,6 @@ type Delivery struct {
 	Priority       int
 	MaxWait        int
 	PickUpTime     int64
-	CookingTime    int
+	CookingTime    int64
 	CookingDetails []props.CookingDetail
 }
